adding documentation and ordering functions in appropriate libraries for afm simulations, calculation and least square optimization routines
